requirements.txt

langchain==0.1.0
anthropic==0.18.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pypdf==4.0.1
python-docx==1.1.0
datasets==2.16.1
rank-bm25==0.2.2
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
redis==5.0.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
tqdm==4.66.1
requests==2.31.0
pytest==8.0.0
pytest-asyncio==0.22.0
matplotlib==3.8.2
seaborn==0.13.2
pandas==2.1.4
plotly==5.18.0
httpx==0.24.1
pytest==8.0.0